+++ pa_dlna/libpulse/__init__.py
@@ -1,7 +1,7 @@
 """An asyncio interface to the Pulseaudio library."""
 
-from .pulselib import (PulseLib, PulseEvent, EventIterator,
-                       PulseLibError, PulseMissingLibError, PulseClosedError,
+from .libpulse import (LibPulse, PulseEvent, EventIterator,
+                       LibPulseError, PulseMissingLibError, PulseClosedError,
                        PulseStateError, PulseOperationError,
                        PulseClosedIteratorError,)
 from .pulseaudio_h import *